refactor adminSemesterUtil and adminClientUtil into one handler file
