+++ src/app/(frontend)/admin/page.tsx
@@ -8,13 +8,13 @@ import React, { useEffect, useState } from 'react'
 import { motion } from 'framer-motion'
 import { mockProjects } from '@/test-config/mocks/Project.mock'
 import { handleLoginButtonClick, getLoggedInUser } from '@/lib/services/user/Handlers'
-import { getAllSemesters } from '@/lib/util/adminSemesterUtils'
+import { getAllSemesters } from '@/lib/services/admin/Handlers'
 import { Semester } from '@/payload-types'
 import MobileAdminView from '@/app/(frontend)/admin/MobileAdminView'
 import { UserCombinedInfo } from '@/types/Collections'
 import { redirect } from 'next/navigation'
 import { UserRole } from '@/types/User'
-import { getAllClients } from '@/lib/util/adminClientUtils'
+import { getAllClients } from '@/lib/services/admin/Handlers'
 import { ProjectDetails } from '@/types/Project'
 import Notification from '@/components/Generic/Notification/Notification'
